+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Launch instances from env.reset()
 - Send commands through env.step()
 - Minor command and logging changes
+- Configurable action/observation ports
 - MCIO_PROTOCOL_VERSION = 2
 
 ## 0.2.0 - 2024-12-22